python-web-apps/flask-online-bookstore/requirements.txt
blinker==1.7.0
click==8.1.7
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==3.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
Pillow==10.1.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
Werkzeug==3.0.1
WTForms==3.1.1
